reorganize(resolveExternalValue)
