--- packages/delegate/src/resolveExternalValue/handleNull.ts
@@ -1,28 +1,0 @@
-import { GraphQLError, locatedError } from 'graphql';
-
-import AggregateError from '@ardatan/aggregate-error';
-
-const reportedErrors: WeakMap<GraphQLError, boolean> = new Map();
-
-export function handleNull(unpathedErrors: Array<GraphQLError>) {
-  if (unpathedErrors.length) {
-    const unreportedErrors: Array<GraphQLError> = [];
-    unpathedErrors.forEach(error => {
-      if (!reportedErrors.has(error)) {
-        unreportedErrors.push(error);
-        reportedErrors.set(error, true);
-      }
-    });
-
-    if (unreportedErrors.length) {
-      if (unreportedErrors.length === 1) {
-        return unreportedErrors[0];
-      }
-
-      const combinedError = new AggregateError(unreportedErrors);
-      return locatedError(combinedError, undefined, unreportedErrors[0].path);
-    }
-  }
-
-  return null;
-}
